add workflow

+++ scripts/setup_kind_cluster.sh
@@ -5,6 +5,7 @@ set -e
 function show_help() {
   echo "Usage: setup_kind_cluster.sh [OPTIONS]"
   echo "Options:"
+  echo "  --ci                 Setup cluster for a CI environment"
   echo "  --reset              Reset the existing kind cluster"
   echo "  --private            Use private repositories. Requires a PAT with read:packages scope."
   echo "  --build-deps         Build dependencies for all charts"
@@ -26,6 +27,11 @@ PROJECT_ROOT=$(dirname "$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)")
 KUBECONFIG_FILE=$PROJECT_ROOT/tracetest.kubeconfig
 ENV_FILE=$PROJECT_ROOT/cluster.env
 HELM_EXTRA_FLAGS=()
+VALUES_FILE=values-kind.yaml
+
+if [[ "$@" == *"--ci"* ]]; then
+  VALUES_FILE=values-kind-ci.yaml
+fi
 
 if [[ "$@" == *"--debug"* ]]; then
   set -x
@@ -68,12 +74,23 @@ else
   printf "\n\e[1mCluster already exists\e[0m\n"
 fi
 
-cat <<EOF > $ENV_FILE
+# the kind cluster might have been created from outside this script, as in CI pipelines.
+# if that's the case, leave kubeconfig alone
+if [[ -f $KUBECONFIG_FILE ]]; then
+  cat <<EOF > $ENV_FILE
 export KUBECONFIG=$KUBECONFIG:$KUBECONFIG_FILE
 kubectl config use-context kind-tracetest
 EOF
 
-source $ENV_FILE
+  source $ENV_FILE
+else
+  printf "\n\e[43m\e[1mCustom Kubeconfig file not detected, use default kubectl config\e[0m\e[0m\n"
+fi
+
+if [[ "$@" == *"--force-setup"* ]]; then
+  printf "\n\e[43m\e[1mForce setup flag detected\e[0m\e[0m\n"
+  SETUP_CLUSTER=true
+fi
 
 if [[ "$SETUP_CLUSTER" == true ]]; then
 
@@ -116,10 +133,10 @@ fi
 printf "\n\e[42m\e[1mStarting Tracetest OnPrem components installation\e[0m\e[0m\n"
 
 printf "\n\e[42m\e[1mInstalling Tracetest dependencies\e[0m\e[0m\n"
-helm upgrade --install ttdeps $PROJECT_ROOT/charts/tracetest-dependencies -f $PROJECT_ROOT/values-kind.yaml "${HELM_EXTRA_FLAGS[@]}"
+helm upgrade --install ttdeps $PROJECT_ROOT/charts/tracetest-dependencies -f $PROJECT_ROOT/$VALUES_FILE "${HELM_EXTRA_FLAGS[@]}"
 
 printf "\n\e[42m\e[1mInstalling Tracetest on-prem\e[0m\e[0m\n"
-helm upgrade --install tt $PROJECT_ROOT/charts/tracetest-onprem -f $PROJECT_ROOT/values-kind.yaml "${HELM_EXTRA_FLAGS[@]}"
+helm upgrade --install tt $PROJECT_ROOT/charts/tracetest-onprem -f $PROJECT_ROOT/$VALUES_FILE "${HELM_EXTRA_FLAGS[@]}"
 
 if [[ "$@" == *"--install-demo"* ]]; then
   printf "\n\e[42m\e[1mInstalling Pokeshop demo\e[0m\e[0m\n"
